requirements.txt
-i https://pypi.org/simple
aioboto3==12.1.0; python_version >= '3.8' and python_version < '4.0'
aiobotocore[boto3]==2.8.0; python_version >= '3.8'
aiohttp==3.9.1; python_version >= '3.8'
aioitertools==0.11.0; python_version >= '3.6'
aioredis==2.0.1; python_version >= '3.6'
aiosignal==1.3.1; python_version >= '3.7'
alembic==1.13.0; python_version >= '3.8'
annotated-types==0.6.0; python_version >= '3.8'
anyio==3.7.1; python_version >= '3.7'
async-timeout==4.0.3; python_version >= '3.7'
asyncio==3.4.3
asyncpg==0.29.0; python_full_version >= '3.8.0'
attrs==23.1.0; python_version >= '3.7'
boto3==1.33.1; python_version >= '3.7'
botocore==1.33.1; python_version >= '3.7'
click==8.1.7; python_version >= '3.7'
fastapi==0.104.1; python_version >= '3.8'
frozenlist==1.4.0; python_version >= '3.8'
greenlet==3.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
idna==3.6; python_version >= '3.5'
jmespath==1.0.1; python_version >= '3.7'
mako==1.3.0; python_version >= '3.8'
markupsafe==2.1.3; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
pydantic==2.5.2; python_version >= '3.7'
pydantic-core==2.14.5; python_version >= '3.7'
pydantic-settings==2.1.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.0; python_version >= '3.8'
s3transfer==0.8.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.0; python_version >= '3.7'
sqlalchemy==2.0.23; python_version >= '3.7'
starlette==0.27.0; python_version >= '3.7'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==2.0.7; python_version >= '3.10'
uvicorn==0.24.0; python_version >= '3.8'
wrapt==1.16.0; python_version >= '3.6'
yarl==1.9.4; python_version >= '3.7'
